Move add_constraint to Pyomo-utils
