+++ model/common/pyomo_utils.py
@@ -86,6 +86,18 @@ class RegionalInitConstraint(GeneralConstraint):
         return Constraint(m.regions, rule=self.rule)
 
 
+def add_constraint(m, constraint, name=None):
+    """Adds a constraint to the model
+
+    It first generates a unique name, then adds
+    the constraint using this new name
+    """
+    n = len(list(m.component_objects()))
+    name = f"constraint_{n}" if name is None else f"constraint_{name}"
+    m.add_component(name, constraint)
+    return name
+
+
 ####### Get all variables of a model
